+++ src/test/java/mil/nga/geopackage/GeoPackageImportTest.java
@@ -122,4 +122,17 @@ public class GeoPackageImportTest extends ImportGeoPackageTestCase {
 
 	}
 
+	/**
+	 * Test user daos
+	 * 
+	 * @throws SQLException
+	 *             upon error
+	 */
+	@Test
+	public void testUserDao() throws SQLException {
+
+		GeoPackageTestUtils.testUserDao(geoPackage);
+
+	}
+
 }