+++ src/test/java/mil/nga/geopackage/GeoPackageTestUtils.java
@@ -1,5 +1,7 @@
 package mil.nga.geopackage;
 
+import static org.junit.Assert.fail;
+
 import java.io.File;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -10,11 +12,13 @@ import java.util.Map;
 import java.util.Set;
 
 import junit.framework.TestCase;
+import mil.nga.geopackage.attributes.AttributesDao;
 import mil.nga.geopackage.contents.Contents;
 import mil.nga.geopackage.contents.ContentsDao;
 import mil.nga.geopackage.contents.ContentsDataType;
 import mil.nga.geopackage.db.GeoPackageDataType;
 import mil.nga.geopackage.db.TableColumnKey;
+import mil.nga.geopackage.db.table.TableInfo;
 import mil.nga.geopackage.extension.coverage.CoverageData;
 import mil.nga.geopackage.features.columns.GeometryColumns;
 import mil.nga.geopackage.features.columns.GeometryColumnsDao;
@@ -29,6 +33,8 @@ import mil.nga.geopackage.tiles.matrix.TileMatrixDao;
 import mil.nga.geopackage.tiles.matrixset.TileMatrixSet;
 import mil.nga.geopackage.tiles.matrixset.TileMatrixSetDao;
 import mil.nga.geopackage.tiles.user.TileDao;
+import mil.nga.geopackage.user.UserDao;
+import mil.nga.geopackage.user.custom.UserCustomDao;
 import mil.nga.proj.Projection;
 import mil.nga.proj.ProjectionConstants;
 import mil.nga.proj.ProjectionFactory;
@@ -719,6 +725,54 @@ public class GeoPackageTestUtils {
 			}
 
 			previousType = type;
+		}
+
+	}
+
+	/**
+	 * Test user daos
+	 * 
+	 * @param geoPackage
+	 *            GeoPackage
+	 * @throws SQLException
+	 *             upon error
+	 */
+	public static void testUserDao(GeoPackage geoPackage) throws SQLException {
+
+		List<String> tables = geoPackage.getTables();
+		for (String table : tables) {
+			UserDao<?, ?, ?, ?> dao = geoPackage.getUserDao(table);
+			TestCase.assertNotNull(dao);
+			ContentsDataType dataType = geoPackage.getTableCoreDataType(table);
+			TableInfo tableInfo = TableInfo.info(geoPackage.getConnection(),
+					table);
+			TestCase.assertEquals(tableInfo.numColumns(), dao.columnCount());
+			Contents contents = dao.getContents();
+			TestCase.assertNotNull(contents);
+			if (dataType == null) {
+				TestCase.assertTrue(dao instanceof UserCustomDao);
+				TestCase.assertNotNull(contents.getDataTypeName());
+			} else {
+				switch (dataType) {
+				case ATTRIBUTES:
+					TestCase.assertTrue(dao instanceof AttributesDao);
+					TestCase.assertEquals(ContentsDataType.ATTRIBUTES,
+							contents.getDataType());
+					break;
+				case FEATURES:
+					TestCase.assertTrue(dao instanceof FeatureDao);
+					TestCase.assertEquals(ContentsDataType.FEATURES,
+							contents.getDataType());
+					break;
+				case TILES:
+					TestCase.assertTrue(dao instanceof TileDao);
+					TestCase.assertEquals(ContentsDataType.TILES,
+							contents.getDataType());
+					break;
+				default:
+					fail("Unsupported data type: " + dataType);
+				}
+			}
 		}
 
 	}